20260512 @Mookse - created `tools.mjs` as extensible module - created campaign skills drafts   - campaignActivate.json   - campaignClosejson   - campaignConnect.json   - campaignInitialize.json - created functions   - campaignActivate.json   - campaignClosejson   - campaignConnect.json   - campaignInitialize.json
