+++ inc/json-schemas/openai/functions/campaignConnect.json
@@ -1,0 +1,64 @@
+{
+  "description": "Generates a report for a CONNECT event from the intelligence; identifies `cid` (campaign instance id) to be used for subsequent session calls in the same campaign process",
+  "name": "campaignConnect",
+  "parameters": {
+    "additionalProperties": false,
+    "type": "object",
+    "properties": {
+      "cid": {
+        "description": "The unique identifier of the campaign instance",
+        "format": "uuid",
+        "type": "string"
+      },
+      "report": {
+        "additionalProperties": false,
+        "description": "My interpretation of our successful CONNECT events, including the detailed affect vector and any new distress flags, hooks and mood",
+        "type": "object",
+        "properties": {
+          "affect_vector": {
+            "additionalProperties": false,
+            "description": "The improvement of emotional connection over this conversation",
+            "type": "object",
+            "properties": {
+              "label": {
+                "description": "The label/description of the affect vector, ex., 'alone -> felt understood'",
+                "type": "string"
+              },
+              "evidence": {
+                "description": "The evidence for the affect vector, such as 'the visitor seems more engaged and responsive than before'",
+                "type": "string"
+              },
+              "intensity": {
+                "description": "The intensity of the affect vector",
+                "enum": ["low", "medium", "high"],
+                "type": "string"
+              }
+            },
+            "required": ["label", "evidence", "intensity"]
+          },
+          "distress_flags": {
+            "description": "Any noticed distress flags set by the visitor contact. Content that seems difficult for human; comma-delimited",
+            "type": "string"
+          },
+          "hook": {
+            "description": "What's the most compelling argument for the visitor so far?",
+            "type": "string"
+          },
+          "mood": {
+            "description": "What's the general mood of the visitor at this point?",
+            "type": "string"
+          },
+          "stage": {
+            "const": "connected",
+            "description": "The stage of the visitor in the campaign process",
+            "enum": ["initialized", "open", "connected", "activated", "closed"],
+            "type": "string"
+          }
+        },
+        "required": ["affect_vector", "distress_flags", "hook", "mood", "stage"]
+      }
+    },
+    "required": ["cid", "report"]
+  },
+  "strict": true
+}
